SCRUM-228 fix: refactoring
