+++ src/widgets/myPayments/lib/formatSubscriptionType.ts
@@ -1,0 +1,14 @@
+import { SelectedSubscriptionType } from '@/src/entities/subscription/types'
+
+export const formatSubscriptionType = (type: SelectedSubscriptionType): string => {
+  switch (type) {
+    case 'DAY':
+      return '1 day'
+    case 'WEEKLY':
+      return '7 days'
+    case 'MONTHLY':
+      return '1 month'
+    default:
+      return type
+  }
+}
